scaffolded postgres database, enum mapping still missing (info on discord)

+++ Acorn.BL/Models/Account.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using Acorn.BL.Helpers;
 
 namespace Acorn.BL.Models
@@ -7,7 +6,7 @@ namespace Acorn.BL.Models
     public partial class Account
     {
         public long AccountId { get; set; }
-        public long BotId { get; set; }
+        public int BotId { get; set; }
         public string Login { get; set; }
         public string Password { get; set; }
         public DateTime BirthDate { get; set; }
